+++ infrastructures/PhpSecLib/Exception/WrongLibraryAPIException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * East Paas.
  *
@@ -21,23 +23,18 @@
  * @author      Richard Déloge <richard@teknoo.software>
  */
 
-declare(strict_types=1);
+namespace Teknoo\East\Paas\Infrastructures\PhpSecLib\Exception;
 
-namespace Teknoo\East\Paas\Infrastructures\Symfony\Messenger\Message;
-
-use Stringable;
-use Teknoo\East\Paas\Contracts\Message\MessageInterface;
-use Teknoo\Immutable\ImmutableInterface;
+use RuntimeException;
 
 /**
- * Message object, for Symfony Messenger, about any job's event to persist in the database
+ * API instance injected in PHPSecLib integration is not compliant with East PaaS
  *
  * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
  * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
  * @license     http://teknoo.software/license/mit         MIT License
  * @author      Richard Déloge <richard@teknoo.software>
  */
-class HistorySent implements MessageInterface, Stringable, ImmutableInterface
+class WrongLibraryAPIException extends RuntimeException
 {
-    use MessageTrait;
 }